FE: add react-router-dom, hashrouter, hook, Movie detail
