+++ frontend/package.json
@@ -11,9 +11,9 @@
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "axios": "^0.27.2",
+    "bootstrap": "5.0.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
-    "bootstrap": "5.0.2",
     "react-dom": "^18.2.0",
     "react-modal": "^3.15.1",
     "react-router-dom": "^6.4.1",
